--- a/bench/ets_hash_ring_bench.exs
+++ b/bench/ets_hash_ring_bench.exs
@@ -2,47 +2,70 @@ defmodule ETSHashRingBench do
   use Benchfella
   alias ExHashRing.HashRing.ETS, as: Ring
 
-  @nodes [
-    "hash-ring-1-1",
-    "hash-ring-1-2",
-    "hash-ring-1-3",
-    "hash-ring-1-4"
-  ]
-  @replicas 512
   @name HashRingBench.ETSRing
+  @nodes ["hash-ring-1-1", "hash-ring-1-2", "hash-ring-1-3", "hash-ring-1-4"]
+  @replicas 512
+  @overrides %{"1234254543" => 1}
 
   setup_all do
     Ring.Config.start_link()
     {:ok, nil}
   end
 
-  before_each_bench _ do
-    {:ok, _pid} = Ring.start_link(@name, nodes: @nodes, num_replicas: @replicas, named: true)
-    {:ok, @name}
-  end
-
   after_each_bench _ do
     GenServer.stop(@name)
   end
 
-  bench "find node" do
-    Ring.find_node(bench_context, "1234254543")
+  bench "find_node", ring: new_ring() do
+    Ring.find_node(ring, "1234254543")
     :ok
   end
 
-  bench "find nodes (1)" do
-    Ring.find_nodes(bench_context, "1234254543", 2)
+  bench "find_nodes(num: 2)", ring: new_ring() do
+    Ring.find_nodes(ring, "1234254543", 2)
     :ok
   end
 
-  bench "find nodes (2)" do
-    Ring.find_nodes(bench_context, "1234254543", 3)
+  bench "find_nodes(num: 3)", ring: new_ring() do
+    Ring.find_nodes(ring, "1234254543", 3)
     :ok
   end
 
-  bench "regenerate ring & gc" do
-    Ring.set_nodes(bench_context, @nodes)
-    Ring.force_gc(bench_context)
+  bench "find_node [override, match=true]", ring: new_ring(@overrides) do
+    Ring.find_node(ring, "1234254543")
+    :ok
+  end
+
+  bench "find_node [override, match=false]", ring: new_ring(@overrides) do
+    Ring.find_node(ring, "0")
+    :ok
+  end
+
+  bench "find_nodes(num: 2) [override]", ring: new_ring(@overrides) do
+    Ring.find_nodes(ring, "1234254543", 2)
+    :ok
+  end
+
+  bench "find_nodes(num: 3) [override]", ring: new_ring(@overrides) do
+    Ring.find_nodes(ring, "1234254543", 3)
+    :ok
+  end
+
+  bench "regenerate ring & gc", ring: new_ring() do
+    Ring.set_nodes(ring, @nodes)
+    Ring.force_gc(ring)
     nil
+  end
+
+  defp new_ring(overrides \\ %{}) do
+    {:ok, _} =
+      Ring.start_link(@name,
+        nodes: @nodes,
+        num_replicas: @replicas,
+        overrides: overrides,
+        named: true
+      )
+
+    @name
   end
 end