… test/hash_ring_test.exs