- allows clients to provide specific key overrides for a ring.
- performance is generally improved.
- effectively zero performance impact when using small number of overrides (relative to v3.0.0).

|                        |        before |         after |        + override |
| :--------------------- | ------------: | ------------: | ----------------: |
| `find_node`            |    0.73 µs/op |    0.73 µs/op | 0.17 - 0.73 µs/op |
| `find_nodes(num: 2)`   |    1.61 µs/op |    1.55 µs/op | 0.79 - 1.08 µs/op |
| `find_nodes(num: 3)`   |    2.17 µs/op |    2.12 µs/op | 1.63 - 2.18 µs/op |
| `regenerate ring & gc` | 2974.22 µs/op | 3064.08 µs/op |     3008.48 µs/op |